backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
anthropic==0.45.0
playwright==1.50.0
aiosqlite==0.20.0
python-dotenv==1.0.1
mss==9.0.2
pyautogui==0.9.54
pydantic==2.10.4
